app/schemas/admin.py
from __future__ import annotations

from pydantic import BaseModel


class AdminLoginIn(BaseModel):
    username: str
    password: str


class AdminTokenOut(BaseModel):
    access_token: str
    token_type: str = "bearer"

